+++ src/main/kotlin/dev/aaronhowser/mods/irregular_implements/item/BiomeCapsuleItem.kt
@@ -49,7 +49,11 @@ class BiomeCapsuleItem(properties: Properties) : Item(properties) {
 		val DEFAULT_PROPERTIES: Properties = Properties().stacksTo(1)
 
 		fun getItemColor(stack: ItemStack, tintIndex: Int): Int {
-			val foliageColor = stack.get(ModDataComponents.BIOME)?.value()?.foliageColor ?: 0xFFFFFFFF.toInt()
+			val foliageColor = stack.get(ModDataComponents.BIOME_POINTS)
+				?.biome
+				?.value()
+				?.foliageColor
+				?: 0xFFFFFFFF.toInt()
 
 			return (foliageColor or 0xFF000000.toInt())
 		}